Login: Move helpers in utils file
